✨ First working version
The rate-limit is relatively low, so we try to make use of DokuWiki's
instruction caching functionality. However we may have to provide a
button or something to clear the cache manually to prevent confusion on
the customers end.
On the other hand, the rate limit seems to be 250 Hits, reset every 15
minutes, which should be enough to use ~~NOCACHE~~ if necessary.

Reducing the transmitted fields with filters is prerequisite to have
that rate-limit of 250, otherwise we would only have 100 Hits.

ToDo: Error handling, Styling

PCC-52