don't use keep-alive to communicate with the API
Vimeo seems to use a new cache in front of the API. It responds without
a Content-Length response on a keep-alive connection and does not use
chunked encoding either. I'm unsure why. Doing the same request with
curl uses chunked encoding correctly. Disabling keep-alive is a good
workaround for now.